# kcdf-api-backend/public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KCDF Parents Platform \u2014 API",
        "description": "REST API for the KCDF Parents platform. Serves Parent App and Admin Portal.",
        "version": "1.0.0",
        "contact": {
            "name": "KCDF Support",
            "email": "support@kcdf.org"
        },
        "license": {
            "name": "Proprietary"
        }
    },
    "servers": [
       
        {
            "url": "https://api.kcdfindia.com",
            "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Bearer token for authentication"
            }
        },
...
    },
    "paths": {
...
        "/api/v1/activity-logs": {
            "get": {
                "operationId": "listActivityLogs",
                "tags": [
                    "Activity Logs"
                ],
                "summary": "List activity logs",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "actor_profile_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "entity_type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "action",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity logs retrieved"
                    }
                }
            }
        }
    }
}